Making the branch selector it's own directive moving tests adding $provide.factory where needed to mock it
