+++ client/directives/environment/components/branchSelector/branchSelectorView.jade
@@ -1,0 +1,20 @@
+.label-col Initial Branch
+.input-col
+  fancy-select(
+    ng-disabled = "branchFetching"
+    ng-required = "true"
+    value = "state.branch"
+    track-by = "attrs.name"
+    spinner-flag = "branchFetching"
+  )
+    li.list-item.popover-list-item.small Select initial branch
+    fancy-option(
+      ng-repeat = "branch in $parent.state.repo.branches.models"
+      value = "branch"
+    )
+      svg.iconnables.icons-branch
+        use(
+          xlink:href = "#icons-branch"
+        )
+      | {{ $parent.branch.attrs.name }}
+  small.small We’ll build the first container with the code on this branch. We recommend choosing a stable branch.
