+++ test/unit/environment/repositoryFormDirective.unit.js
@@ -133,201 +133,4 @@ describe('repositoryFormDirective'.bold.underline.blue, function () {
       $rootScope.$destroy();
     });
   });
-  describe('starting up', function () {
-
-    it('for the first time', function () {
-
-      var scope = {
-        state: {}
-      };
-      setup(scope);
-      $scope.$digest();
-      $elScope = ctx.element.isolateScope();
-      $scope.$digest();
-
-      $scope.state.repo = ctx.repo1;
-      $scope.$digest();
-      $scope.state.acv = ctx.acv;
-      $scope.$digest();
-      $rootScope.$apply();
-
-      sinon.assert.called(ctx.repo1.newBranch);
-      expect($scope.state.branch, 'branch').to.be.ok;
-      $scope.$digest();
-      sinon.assert.called(ctx.repo1.fakeBranch.fetch);
-
-      sinon.assert.called(ctx.repo1.branches.fetch);
-      $scope.$digest();
-      expect($scope.state.branch.attrs.name, 'branch name').to.equal('master');
-      $scope.$digest();
-      sinon.assert.notCalled(ctx.acv.update);
-      sinon.assert.notCalled(ctx.acv.setState);
-
-      // Now make a change
-      var newBranch = ctx.repo1.branches.models[0];
-      $elScope.state.branch = newBranch;
-      $scope.$digest();
-      var newState = {
-        branch: newBranch.attrs.name,
-        commit: newBranch.attrs.commit.sha
-      };
-      sinon.assert.calledWith(ctx.acv.setState, newState);
-      sinon.assert.calledWith(ctx.acv.update, newState);
-      $scope.$digest();
-      sinon.assert.calledOnce(ctx.acv.resetState);
-      $rootScope.$destroy();
-    });
-    it('coming back to the page', function () {
-      setup({
-        state: {}
-      });
-      $scope.state.acv = ctx.acv;
-      $scope.state.repo = ctx.repo1;
-      $scope.state.branch = {
-        attrs: apiMocks.branches.bitcoinRepoBranches[0]
-      };
-
-      ctx.repo1.branches.fetch();
-      ctx.repo1.branches.fetch.reset();
-      $scope.$digest();
-      $elScope = ctx.element.isolateScope();
-      $scope.$digest();
-
-      sinon.assert.notCalled(ctx.repo1.newBranch);
-      expect($scope.state.branch, 'branch').to.be.ok;
-      $scope.$digest();
-      sinon.assert.notCalled(ctx.repo1.branches.fetch);
-      sinon.assert.notCalled(ctx.repo1.fakeBranch.fetch);
-      $scope.$digest();
-
-      sinon.assert.notCalled(ctx.acv.update);
-      sinon.assert.notCalled(ctx.acv.setState);
-
-      // Now make a change
-      var newBranch = ctx.repo1.branches.models[1];
-      $elScope.state.branch = newBranch;
-      $scope.$digest();
-      var newState = {
-        branch: newBranch.attrs.name,
-        commit: newBranch.attrs.commit.sha
-      };
-      sinon.assert.calledWithMatch($scope.state.acv.setState, newState);
-      sinon.assert.calledWithMatch($scope.state.acv.update, newState);
-      $scope.$digest();
-      sinon.assert.calledOnce($scope.state.acv.resetState);
-      $rootScope.$destroy();
-    });
-    it('should fetch the branches, but not fetch the main branch ', function () {
-
-      var scope = {
-        state: {
-          branch: {
-            attrs: {
-              name: 'master'
-            }
-          }
-        }
-      };
-      setup(scope);
-      $scope.$digest();
-      $elScope = ctx.element.isolateScope();
-      $scope.$digest();
-
-      $scope.state.repo = ctx.repo1;
-      $scope.$digest();
-      $scope.state.acv = ctx.acv;
-      $scope.$digest();
-      $rootScope.$apply();
-
-
-      sinon.assert.notCalled(ctx.repo1.newBranch);
-      expect($scope.state.branch, 'branch').to.be.ok;
-      $scope.$digest();
-      sinon.assert.notCalled(ctx.repo1.fakeBranch.fetch);
-      $scope.$digest();
-      $scope.$digest();
-
-      sinon.assert.called(ctx.repo1.branches.fetch);
-      $scope.$digest();
-      expect($scope.state.branch.attrs.name, 'branch name').to.equal('master');
-      $scope.$digest();
-    });
-
-    it('with a loadingPromise target', function () {
-      setup({
-        state: {},
-        loadingPromisesTarget: 'helloEverybody'
-      });
-      var loadingStub = sinon.spy(loadingPromises, 'add');
-      $scope.state.acv = ctx.acv;
-      $scope.state.repo = ctx.repo1;
-      $scope.state.branch = {
-        attrs: apiMocks.branches.bitcoinRepoBranches[0]
-      };
-
-      ctx.repo1.branches.fetch();
-      ctx.repo1.branches.fetch.reset();
-      $scope.$digest();
-      $elScope = ctx.element.isolateScope();
-      $scope.$digest();
-
-      sinon.assert.notCalled(ctx.repo1.branches.fetch);
-      $scope.$digest();
-      sinon.assert.notCalled(ctx.acv.update);
-      sinon.assert.notCalled(ctx.acv.setState);
-
-      // Now make a change
-      var newBranch = ctx.repo1.branches.models[1];
-      $elScope.state.branch = newBranch;
-      $scope.$digest();
-
-      loadingPromises.finished();
-      $scope.$digest();
-      $scope.$digest();
-
-      var newState = {
-        branch: newBranch.attrs.name,
-        commit: newBranch.attrs.commit.sha
-      };
-      sinon.assert.calledWithMatch($scope.state.acv.update, newState);
-      sinon.assert.called(loadingStub);
-
-      $scope.$digest();
-      sinon.assert.calledOnce($scope.state.acv.resetState);
-      $rootScope.$destroy();
-    });
-  });
-
-  describe('errors', function () {
-    it('failing on update', function () {
-      var error = new Error('sadasdasd');
-      setup({
-        state: {}
-      }, error);
-      $scope.state.acv = ctx.acv;
-      $scope.state.repo = ctx.repo1;
-      $scope.state.branch = {
-        attrs: apiMocks.branches.bitcoinRepoBranches[0]
-      };
-      ctx.repo1.branches.fetch();
-      ctx.repo1.branches.fetch.reset();
-      $scope.$digest();
-      $elScope = ctx.element.isolateScope();
-      $scope.$digest();
-      // Now make a change
-      var newBranch = ctx.repo1.branches.models[1];
-      $elScope.state.branch = newBranch;
-      $scope.$digest();
-      $scope.$digest();
-      var newState = {
-        branch: newBranch.attrs.name,
-        commit: newBranch.attrs.commit.sha
-      };
-      sinon.assert.calledWithMatch($scope.state.acv.setState, newState);
-      sinon.assert.calledWithMatch($scope.state.acv.update, newState);
-      sinon.assert.calledWith(ctx.errsMock.handler, error);
-      $scope.$digest();
-      sinon.assert.calledOnce($scope.state.acv.resetState);
-    });
-  });
 });
